Update
Updated the project to use bootstrap 5 and a new theme that gives it a futuristic neon look. I have also added razor views for the login and app pages as well as adding the app controller.

My Next move is going to be looking at authenticating a new user. I will create a login form that interfaces with the Spotify OAuth API